Remove witness tx struct (#1457)
### Description

We remove the witness tx struct.

We can wait for the lo-hi refactor ready then for this PR.

### Issue Link

Part of #1391

### Type of change

New feature (non-breaking change which adds functionality)

### Contents

- Remove zkevm-circuits/src/witness/tx.rs entirely
- Remove eth_block field from block
- Acquire txID ASAP so we don't need a Optional<tx_id>

+++ zkevm-circuits/src/evm_circuit/execution/balance.rs
@@ -122,7 +122,7 @@ impl<F: Field> ExecutionGadget<F> for BalanceGadget<F> {
             .assign_h160(region, offset, address.to_address())?;
 
         self.tx_id
-            .assign(region, offset, Value::known(F::from(tx.id as u64)))?;
+            .assign(region, offset, Value::known(F::from(tx.id)))?;
 
         self.reversion_info.assign(
             region,
