+++ zkevm-circuits/src/evm_circuit/execution/sstore.rs
@@ -168,7 +168,7 @@ impl<F: Field> ExecutionGadget<F> for SstoreGadget<F> {
         self.same_context.assign_exec_step(region, offset, step)?;
 
         self.tx_id
-            .assign(region, offset, Value::known(F::from(tx.id as u64)))?;
+            .assign(region, offset, Value::known(F::from(tx.id)))?;
         self.is_static
             .assign(region, offset, Value::known(F::from(call.is_static as u64)))?;
         self.reversion_info.assign(
